--- a/extra_foam/gui/misc_widgets/aesthetics.py
+++ b/extra_foam/gui/misc_widgets/aesthetics.py
@@ -10,8 +10,8 @@ All rights reserved.
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QBrush, QColor, QPen, QPalette
 
-from ..pyqtgraph import ColorMap
-from ..pyqtgraph.graphicsItems.GradientEditorItem import Gradients
+from ..pyqtgraph_old import ColorMap
+from ..pyqtgraph_old.graphicsItems.GradientEditorItem import Gradients
 
 from ...config import config