--- a/extra_foam/gui/plot_widgets/__init__.py
+++ b/extra_foam/gui/plot_widgets/__init__.py
@@ -2,4 +2,3 @@ from .image_items import RectROI, RingItem
 from .image_view_base import ImageViewF, TimedImageViewF
 from .image_views import ImageAnalysis, RoiImageView
 from .plot_widget_base import HistMixin, PlotWidgetF, TimedPlotWidgetF
-from .graphics_widgets import Crosshair