--- a/extra_foam/gui/plot_widgets/graphics_widgets.py
+++ b/extra_foam/gui/plot_widgets/graphics_widgets.py
@@ -20,9 +20,9 @@ from PyQt5.QtWidgets import (
     QGraphicsEllipseItem
 )
 
-from .. import pyqtgraph as pg
-from ..pyqtgraph import Point
-from ..pyqtgraph import functions as fn
+from .. import pyqtgraph_old as pg
+from ..pyqtgraph_old import Point
+from ..pyqtgraph_old import functions as fn
 from ..plot_widgets.plot_items import CurvePlotItem
 from ..misc_widgets import FColor