--- a/extra_foam/gui/plot_widgets/graphics_widgets.py
+++ b/extra_foam/gui/plot_widgets/graphics_widgets.py
@@ -27,35 +27,6 @@ from ..plot_widgets.plot_items import FoamPlotItem, FoamPlotDataItem, CurvePlotI
 from ..misc_widgets import FColor
 
 
-class Crosshair(QGraphicsItemGroup):
-    _HALF_LENGTH = 40
-    _HALF_RADIUS = 2 * _HALF_LENGTH
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        crosshair_pen = FColor.mkPen("r", width=3)
-        circle_pen = FColor.mkPen("k", width=2)
-
-        horiz_crosshair = QGraphicsLineItem()
-        vertical_crosshair = QGraphicsLineItem()
-        circle = QGraphicsEllipseItem()
-
-        horiz_crosshair.setPen(crosshair_pen)
-        vertical_crosshair.setPen(crosshair_pen)
-        circle.setPen(circle_pen)
-
-        self.addToGroup(horiz_crosshair)
-        self.addToGroup(vertical_crosshair)
-        self.addToGroup(circle)
-
-        self.setPos(0, 0)
-        horiz_crosshair.setLine(-self._HALF_LENGTH, 0, self._HALF_LENGTH, 0)
-        vertical_crosshair.setLine(0, -self._HALF_LENGTH, 0, self._HALF_LENGTH)
-        circle.setRect(-self._HALF_RADIUS, -self._HALF_RADIUS,
-                       2 * self._HALF_RADIUS, 2 * self._HALF_RADIUS)
-
-
 class HistogramLUTItem(pg.GraphicsWidget):
     """GraphicsWidget for adjusting the display of an image.