--- a/extra_foam/gui/plot_widgets/image_items.py
+++ b/extra_foam/gui/plot_widgets/image_items.py
@@ -18,9 +18,8 @@ from PyQt5.QtCore import (
     pyqtSignal, pyqtSlot, QPoint, QPointF, QRectF, Qt
 )
 
-from .. import pyqtgraph_old as pg
-from ..pyqtgraph_old import Point
-from ..pyqtgraph_old import functions as fn
+import pyqtgraph as pg
+from pyqtgraph import functions as fn
 
 from ..misc_widgets import FColor
 from ...algorithms import quick_min_max
@@ -41,8 +40,8 @@ class ImageItem(pg.GraphicsObject):
     draw_region_changed_sgn = pyqtSignal(int, int)  # (x, y)
     draw_finished_sgn = pyqtSignal()
 
-    def __init__(self, image=None, parent=None):
-        super().__init__(parent=parent)
+    def __init__(self, image=None):
+        super().__init__()
 
         self._image = None   # original image data
         self._qimage = None  # rendered image for display
@@ -191,8 +190,8 @@ class ImageItem(pg.GraphicsObject):
         if o is None or x is None or y is None:
             return
 
-        w = Point(x-o).length()
-        h = Point(y-o).length()
+        w = pg.Point(x-o).length()
+        h = pg.Point(y-o).length()
         if w == 0 or h == 0:
             self._qimage = None
             return
@@ -319,8 +318,8 @@ class ImageItem(pg.GraphicsObject):
         o = self.mapToDevice(QPointF(0, 0))
         x = self.mapToDevice(QPointF(1, 0))
         y = self.mapToDevice(QPointF(0, 1))
-        w = Point(x-o).length()
-        h = Point(y-o).length()
+        w = pg.Point(x-o).length()
+        h = pg.Point(y-o).length()
         if w == 0 or h == 0:
             self._qimage = None
             return
@@ -377,12 +376,12 @@ class MaskItem(pg.GraphicsObject):
     _TRANSPARENT = QColor(0, 0, 0, 0)
     _COLOR_FORMAT = QImage.Format_ARGB32
 
-    def __init__(self, item, parent=None):
+    def __init__(self, item):
         """Initialization.
 
         :param ImageItem item: a reference to the masked image item.
         """
-        super().__init__(parent=parent)
+        super().__init__()
         if not isinstance(item, ImageItem):
             raise TypeError("Input item must be an ImageItem instance.")
 
@@ -560,8 +559,8 @@ class RectROI(pg.ROI):
 
 
 class GeometryItem(pg.GraphicsObject):
-    def __init__(self, pen=None, brush=None, parent=None):
-        super().__init__(parent=parent)
+    def __init__(self, pen=None, brush=None):
+        super().__init__()
 
         if pen is None and brush is None:
             self._pen = FColor.mkPen('b')