--- a/extra_foam/gui/plot_widgets/image_view_base.py
+++ b/extra_foam/gui/plot_widgets/image_view_base.py
@@ -207,7 +207,7 @@ class ImageViewF(QWidget):
         self._image_item.resetTransform()
 
         if scale is not None:
-            self._image_item.scale(*scale)
+            self._image_item.setScaleXY(*scale)
         if pos is not None:
             self._image_item.setPos(*pos)