--- a/extra_foam/gui/pyqtgraph_old/metaarray/__init__.py
+++ b/extra_foam/gui/pyqtgraph_old/metaarray/__init__.py
@@ -1,1 +1,0 @@
-from .MetaArray import *