Previously, EXtra-foam was using a modified internal copy of pyqtgraph. For the sake of ease-of-maintenance, we decided to remove this and use upstream pyqtgraph instead ([currently a commit from master](https://github.com/European-XFEL/EXtra-foam/blob/pyqtgraph/setup.py#L319)). That's all this PR does, functionality is left exactly the same (with the possible exception of the crosshair widget in the azimuthal integration window, which looks different [see 65f8d3c2]).

The main benefit of the internal copy was plotting performance, and in most cases upstream is still slower:
| Plot type | Benchmark data size | Old FPS | New FPS |
|---|---|---|---|
| Scatter | 5000 | 80 | 26 |
| Line | 5000 | 80 | 40 |
| Bar | 300 | 120 | 67 |
| Error bar | 500 | 93 | 60 |
| Image | 1024x1280 | 33 | 40 |

I benchmarked this with the scripts in `benchmarks/gui/` on my laptop, fullscreen window on a 2K monitor. TL;DR: everything but images is way slower. So, should we still do this? IMHO yes, because:
- I do not want to have to maintain our own pyqtgraph...
- I don't think the performance is a deal-breaker, it's rare that I see fullscreen plots with thousands of points.
- Having said that, such large plots are definitely still possible, e.g. plotting something vs time. But here we have still have options:
  - Use a `TimedPlotWidgetF` to only update the plots every second or so.
  - Use dedicated processes for rendering: https://pyqtgraph.readthedocs.io/en/latest/widgets/remotegraphicsview.html
  - And if absolutely necessary, backport the optimizations from the old version.

I would be interested in knowing if there's another reason not to merge I haven't thought of, or cases where plotting (excluding images) was a bottleneck.