@zhujun98 2022-02-10
To be honest, I don't agree here. You may have seen a few use cases in the hutch, but maybe not all of them. `EXtra-foam` is oftentimes used in multiple big screens together. If you add up the slowdown of multiple plots, it makes a huge difference.