@JamesWrigley 2022-02-10
Mmmm interesting, thanks for replying guys, that's very useful to know. I was hoping to avoid backporting the old optimizations, but it does look like the neatest way to get good performance across the board...