@JamesWrigley 2022-02-10
> I spent considerable time to remove a lot of craps (due to historical reason) in `pyqtgraph`. I also removed unneeded options which slow down the performance.
> 
> I don't think you would require too much effort to maintain it. Also, it allows to add new features quickly without wading through the `pyqtgraph` code base.

So, in general I'm not a fan of forking libraries for internal use; firstly because it puts the burden of maintenance (however small) on us, secondly because I don't think it's in the FOSS spirit to not contribute our improvements back to the community. As much effort as I'm sure it took to optimize and clean up pyqtgraph, I would rather those improvements go upstream :)

> Before I left, @philsmt also proposed to make an internal plotting library. Since I would left soon at that time, this did not happen. I would say that it is also a good direction to go and then you can offload the burden from your should and give it to @CammilleCC :)

Oh? I did not know of this :eyes: What was the motivation?