@zhujun98 2022-02-10
> So, in general I'm not a fan of forking libraries for internal use; firstly because it puts the burden of maintenance (however small) on us, secondly because I don't think it's in the FOSS spirit to not contribute our improvements back to the community. As much effort as I'm sure it took to optimize and clean up pyqtgraph, I would rather those improvements go upstream :)

You will have to break a lot of legacy code from others, which is basically mission impossible. `pyqtgraph` trades flexibility for speed because of the broad user community. The job of DA however is to facilitate the experiments at XFEL at much as possible. 

I did make a critical contribute to `pyqtgraph` 
 https://github.com/pyqtgraph/pyqtgraph/issues/1057#issuecomment-731702469 
to improve the line plot speed, otherwise they would be even slower.

> Oh? I did not know of this 👀 What was the motivation?

Because we all know there are a lot of craps in `pyqtgraph`. 

Again, uses would be happy to see more new features and improvements in the future instead of the other way round.