@leguyader 2022-02-11
> To be honest, I don't agree here. You may have seen a few use cases in the hutch, but maybe not all of them. `EXtra-foam` is oftentimes used in multiple big screens together. If you add up the slowdown of multiple plots, it makes a huge difference.

I can confirm also that at SCS we usually have extra-foam showing many plots and the performance quickly degrade as we show more plots. Slower plotting is probably not the way to go.