@zhujun98 2023-03-30
> 

Image | 1024x1280 | 33 | 40
-- | -- | -- | --

`pyqtgraph` uses the grey colormap by default, which is faster than the other colormap because of some shortcut in rendering. This may be the reason of the performance difference. Anyway, I come to announce that [foamgraph](https://github.com/zhujun98/foamgraph.git) is already in a good shape). I finally found some time to replace the `pyqtgraph` "engine" and it is almost done :-) Of course, a lot of new features have been added.

I'd like to draw your attention and, if you had time, please give some feedback @JamesWrigley  @ebadkamil  @CammilleCC  @dgoeries. Another sprint is still required to make it a stable version. For now, I'd like to know which critical plotting features (e.g. plot type, context menu functions) are missing in the old version. Of course, I also have some "crazy" ideas in my mind and [foamgraph](https://github.com/zhujun98/foamgraph.git) may have another evolution soon.

Performance-vise, it should be even faster because the original `ViewBox` is replaced with `Canvas`.