@jo-moeller 2023-03-30
Are image plots with logarithmic colorbar possible?