@zhujun98 2023-03-30
> Are image plots with logarithmic colorbar possible?

@jo-moeller Yes. In `foamgraph` one sets log scale via the context menu of each axis instead of the view box. So by checking the log scale on the axis of the colorbar, the color scale is expected to become logarithmic. But the display of the colorbar itself does not change. I still need to implement it https://github.com/zhujun98/foamgraph/pull/67. But it is not complicated. I was actually wondering why one would like to make the axis of the colorbar logarithmic, so there is a good use case. Thank you!