@j9ac9k 2023-07-10
greetings from pyqtgraph land.  Just chiming in here to see if we can help here.  Since this PR was opened, we've had _significant_ performance improvements.  Some of the recent points we've hit regarding performance improvements.  I'm not sure what version was checked at the start of this PR, I will assume it was 0.12.3 since that was the current version released at the time this PR was created. 

Some notable changes

* (master branch, will be in the next release) - `PlotDataItem.setDownsample()` sees a _significant_ performance improvement. 
* 0.13+ toggle `enableExperimental` and we see a 5-15% performance improvement (varies by platform) on curves
    * We do some horrible things to construct `QPainterPath` w/o serialization/deserialization [see here](https://www.github.com/pyqtgraph/pyqtgraph/pull/2324)
* 0.12.4 We introduced high-performance line plots for _thick_ lines (lines thicker than 1 pixel were a major pain point for some time).

0.12.3 was released at the time this PR was created, but I feel like I should mention that in this version, there was a _significant_ performance improvement made to line plots.  In 0.12.2 a performance improvement was made in our scatter plots, but it largely hasn't seen much update since.  In that same version, we also added optional numba support for some image item related performance...but it's impossible to say if that would be something you would benefit from w/o knowing more.

Anyway, if upstream is still lagging in some way, I'd be happy to take a closer look.